+++ sfAudio/external/headers/minimp3/minimp3_ex.h
@@ -274,7 +274,7 @@ int mp3dec_detect_cb(mp3dec_io_t *io, uint8_t *buf, size_t buf_size)
             filled = MINIMP3_BUF_SIZE;
     }
     int free_format_bytes, frame_size;
-    mp3d_find_frame(buf, filled, &free_format_bytes, &frame_size);
+    mp3d_find_frame(buf, static_cast<int>(filled), &free_format_bytes, &frame_size);
     if (frame_size)
         return 0; /* MAX_FRAME_SYNC_MATCHES consecutive frames found */
     return MP3D_E_USER;
@@ -355,12 +355,12 @@ int mp3dec_load_cb(mp3dec_t *dec, mp3dec_io_t *io, uint8_t *buf, size_t buf_size
                 if (eof)
                     mp3dec_skip_id3v1(buf, &filled);
             }
-            i = mp3d_find_frame(buf + consumed, filled - consumed, &free_format_bytes, &frame_size);
+            i = mp3d_find_frame(buf + consumed, static_cast<int>(filled - consumed), &free_format_bytes, &frame_size);
             consumed += i;
             hdr = buf + consumed;
         } else
         {
-            i = mp3d_find_frame(buf, buf_size, &free_format_bytes, &frame_size);
+            i = mp3d_find_frame(buf, static_cast<int>(buf_size), &free_format_bytes, &frame_size);
             buf      += i;
             buf_size -= i;
             hdr = buf;
@@ -405,7 +405,7 @@ int mp3dec_load_cb(mp3dec_t *dec, mp3dec_io_t *io, uint8_t *buf, size_t buf_size
     } while(1);
     size_t allocated = MINIMP3_MAX_SAMPLES_PER_FRAME*sizeof(mp3d_sample_t);
     if (detected_samples)
-        allocated += detected_samples*sizeof(mp3d_sample_t);
+        allocated += static_cast<size_t>(detected_samples)*sizeof(mp3d_sample_t);
     else
         allocated += (buf_size/frame_info.frame_bytes)*samples*sizeof(mp3d_sample_t);
     info->buffer = (mp3d_sample_t*)malloc(allocated);
@@ -441,11 +441,11 @@ int mp3dec_load_cb(mp3dec_t *dec, mp3dec_io_t *io, uint8_t *buf, size_t buf_size
                 if (eof)
                     mp3dec_skip_id3v1(buf, &filled);
             }
-            samples = mp3dec_decode_frame(dec, buf + consumed, filled - consumed, info->buffer + info->samples, &frame_info);
+            samples = mp3dec_decode_frame(dec, buf + consumed, static_cast<int>(filled - consumed), info->buffer + info->samples, &frame_info);
             consumed += frame_info.frame_bytes;
         } else
         {
-            samples = mp3dec_decode_frame(dec, buf, MINIMP3_MIN(buf_size, (size_t)INT_MAX), info->buffer + info->samples, &frame_info);
+            samples = mp3dec_decode_frame(dec, buf, static_cast<int>(MINIMP3_MIN(buf_size, (size_t)INT_MAX)), info->buffer + info->samples, &frame_info);
             buf      += frame_info.frame_bytes;
             buf_size -= frame_info.frame_bytes;
         }
@@ -468,7 +468,7 @@ int mp3dec_load_cb(mp3dec_t *dec, mp3dec_io_t *io, uint8_t *buf, size_t buf_size
             samples *= frame_info.channels;
             if (to_skip)
             {
-                size_t skip = MINIMP3_MIN(samples, to_skip);
+                int skip = MINIMP3_MIN(samples, to_skip);
                 to_skip -= skip;
                 samples -= skip;
                 memmove(info->buffer, info->buffer + skip, samples*sizeof(mp3d_sample_t));
@@ -485,7 +485,7 @@ int mp3dec_load_cb(mp3dec_t *dec, mp3dec_io_t *io, uint8_t *buf, size_t buf_size
         }
     } while (frame_info.frame_bytes);
     if (detected_samples && info->samples > detected_samples)
-        info->samples = detected_samples; /* cut padding */
+        info->samples = static_cast<size_t>(detected_samples); /* cut padding */
     /* reallocate to normal buffer size */
     if (allocated != info->samples*sizeof(mp3d_sample_t))
     {
@@ -495,7 +495,7 @@ int mp3dec_load_cb(mp3dec_t *dec, mp3dec_io_t *io, uint8_t *buf, size_t buf_size
         info->buffer = alloc_buf;
     }
     if (frames)
-        info->avg_bitrate_kbps = avg_bitrate_kbps/frames;
+        info->avg_bitrate_kbps = static_cast<int>(avg_bitrate_kbps/frames);
     return ret;
 }
 
@@ -513,7 +513,7 @@ int mp3dec_iterate_buf(const uint8_t *buf, size_t buf_size, MP3D_ITERATE_CB call
     do
     {
         int free_format_bytes = 0, frame_size = 0, ret;
-        int i = mp3d_find_frame(buf, buf_size, &free_format_bytes, &frame_size);
+        int i = mp3d_find_frame(buf, static_cast<int>(buf_size), &free_format_bytes, &frame_size);
         buf      += i;
         buf_size -= i;
         if (i && !frame_size)
@@ -572,7 +572,7 @@ int mp3dec_iterate_cb(mp3dec_io_t *io, uint8_t *buf, size_t buf_size, MP3D_ITERA
     do
     {
         int free_format_bytes = 0, frame_size = 0, ret;
-        int i = mp3d_find_frame(buf + consumed, filled - consumed, &free_format_bytes, &frame_size);
+        int i = mp3d_find_frame(buf + consumed, static_cast<int>(filled - consumed), &free_format_bytes, &frame_size);
         if (i && !frame_size)
         {
             consumed += i;
@@ -665,7 +665,7 @@ static int mp3dec_load_index(void *user_data, const uint8_t *frame, int frame_si
     if (!dec->buffer_samples && dec->index.num_frames < 256)
     {   /* for some cutted mp3 frames, bit-reservoir not filled and decoding can't be started from first frames */
         /* try to decode up to 255 first frames till samples starts to decode */
-        dec->buffer_samples = mp3dec_decode_frame(&dec->mp3d, frame, MINIMP3_MIN(buf_size, (size_t)INT_MAX), dec->buffer, info);
+        dec->buffer_samples = mp3dec_decode_frame(&dec->mp3d, frame, static_cast<int>(MINIMP3_MIN(buf_size, (size_t)INT_MAX)), dec->buffer, info);
         dec->samples += dec->buffer_samples*info->channels;
     } else
         dec->samples += hdr_frame_samples(frame)*info->channels;
@@ -755,7 +755,7 @@ seek_zero:
                 return ret;
         } else
         {
-            int ret = mp3dec_iterate_buf(dec->file.buffer + dec->start_offset, dec->file.size - dec->start_offset, mp3dec_load_index, dec);
+            int ret = mp3dec_iterate_buf(dec->file.buffer + dec->start_offset, static_cast<size_t>(dec->file.size - dec->start_offset), mp3dec_load_index, dec);
             if (ret && MP3D_E_USER != ret)
                 return ret;
         }
@@ -821,7 +821,7 @@ seek_zero:
         }
     }
     dec->offset = dec->index.frames[i].offset;
-    dec->to_skip = position - dec->index.frames[i].sample;
+    dec->to_skip = static_cast<int>(position - dec->index.frames[i].sample);
     while ((i + 1) < dec->index.num_frames && !dec->index.frames[i].sample && !dec->index.frames[i + 1].sample)
     {   /* skip not decodable first frames */
         const uint8_t *hdr;
@@ -893,7 +893,7 @@ size_t mp3dec_ex_read_frame(mp3dec_ex_t *dec, mp3d_sample_t **buf, mp3dec_frame_
             dec_buf = dec->file.buffer + dec->input_consumed;
             if (!(dec->input_filled - dec->input_consumed))
                 return 0;
-            dec->buffer_samples = mp3dec_decode_frame(&dec->mp3d, dec_buf, dec->input_filled - dec->input_consumed, dec->buffer, frame_info);
+            dec->buffer_samples = mp3dec_decode_frame(&dec->mp3d, dec_buf, static_cast<int>(dec->input_filled - dec->input_consumed), dec->buffer, frame_info);
             dec->input_consumed += frame_info->frame_bytes;
         } else
         {
@@ -901,7 +901,7 @@ size_t mp3dec_ex_read_frame(mp3dec_ex_t *dec, mp3d_sample_t **buf, mp3dec_frame_
             uint64_t buf_size = end_offset - dec->offset;
             if (!buf_size)
                 return 0;
-            dec->buffer_samples = mp3dec_decode_frame(&dec->mp3d, dec_buf, MINIMP3_MIN(buf_size, (uint64_t)INT_MAX), dec->buffer, frame_info);
+            dec->buffer_samples = mp3dec_decode_frame(&dec->mp3d, dec_buf, static_cast<int>(MINIMP3_MIN(buf_size, (uint64_t)INT_MAX)), dec->buffer, frame_info);
         }
         dec->buffer_consumed = 0;
         if (dec->info.hz != frame_info->hz || dec->info.layer != frame_info->layer)
@@ -915,7 +915,7 @@ return_e_decode:
             dec->buffer_samples *= frame_info->channels;
             if (dec->to_skip)
             {
-                size_t skip = MINIMP3_MIN(dec->buffer_samples, dec->to_skip);
+                int skip = MINIMP3_MIN(dec->buffer_samples, dec->to_skip);
                 dec->buffer_consumed += skip;
                 dec->to_skip -= skip;
             }
@@ -939,11 +939,11 @@ return_e_decode:
     if (dec->detected_samples)
     {   /* count decoded samples to properly cut padding */
         if (dec->cur_sample + out_samples >= dec->detected_samples)
-            out_samples = dec->detected_samples - dec->cur_sample;
+            out_samples = static_cast<size_t>(dec->detected_samples - dec->cur_sample);
     }
     dec->cur_sample += out_samples;
     *buf = dec->buffer + dec->buffer_consumed;
-    dec->buffer_consumed += out_samples;
+    dec->buffer_consumed += static_cast<int>(out_samples);
     return out_samples;
 }
 
@@ -1176,12 +1176,12 @@ static int mp3dec_open_file_h(HANDLE file, mp3dec_map_info_t *map_info)
     s.LowPart = GetFileSize(file, (DWORD*)&s.HighPart);
     if (s.LowPart == INVALID_FILE_SIZE && GetLastError() != NO_ERROR)
         goto error;
-    map_info->size = s.QuadPart;
+    map_info->size = static_cast<size_t>(s.QuadPart);
 
     mapping = CreateFileMapping(file, NULL, PAGE_READONLY, 0, 0, NULL);
     if (!mapping)
         goto error;
-    map_info->buffer = (const uint8_t*)MapViewOfFile(mapping, FILE_MAP_READ, 0, 0, s.QuadPart);
+    map_info->buffer = (const uint8_t*)MapViewOfFile(mapping, FILE_MAP_READ, 0, 0, static_cast<size_t>(s.QuadPart));
     CloseHandle(mapping);
     if (!map_info->buffer)
         goto error;
