+++ sfAudio/src/System/FileInputStream.cpp
@@ -83,7 +83,7 @@ uint64_t FileInputStream::read(void *data, uint64_t size)
 uint64_t FileInputStream::seek(uint64_t position)
 {
     if (m_file) {
-        if (std::fseek(m_file, static_cast<std::size_t>(position), SEEK_SET))
+        if (std::fseek(m_file, static_cast<long>(position), SEEK_SET))
             return -1;
 
         return tell();
